CDK-848: Add plugin pages generated by maven.
These are generated using the maven-genearted content as an included
HTML blob. The README explains how to add the maven content. The content
should be generated using the empty-maven-skin.

Maven content is copied by the _plugins/copy_from_include.rb plugin. It
copies the contents files, referenced on per line. The files should be
under _includes/. If the files can't be found, then a warning is sent to
stderr and the rendered page has an error message. This method allows
copying content without also interpreting it as markdown and allows
jekyll serve to run even if the files are missing.

Adds redirects for old maven mojo docs locations.