Adding maven plugin pages.
These are generated using the maven-genearted content as an included
HTML blob. The README explains how to depend on the maven content. The
content should be generated using the empty-maven-skin.

For the links from maven to work without modification, this removes the
"pretty" permalink option. Links throughout the other content must be
updated.